support custom config file

+++ package/toolmate/utils/config_tools.py
@@ -41,20 +41,28 @@ if os.path.isdir(storageDir):
         shutil.rmtree(shortcut_dir, ignore_errors=True)
 
         # check if config backup is available
-        backupFile = os.path.join(storageDir, "config_lite_backup.py" if config.isLite else "config_backup.py")
-        if os.path.isfile(backupFile):
-            restore_backup = yes_no_dialog(
-                title="Configuration Backup Found",
-                text=f"Do you want to use the following backup?\n{backupFile}"
-            ).run()
-            if restore_backup:
-                try:
-                    loadConfig(backupFile)
-                    shutil.copy(backupFile, configFile)
-                    print("Configuration backup restored!")
-                    #config.restartApp()
-                except:
-                    print("Failed to restore backup!")
+        if hasattr(config, "custom_config") and config.custom_config and os.path.isfile(config.custom_config):
+            try:
+                loadConfig(config.custom_config)
+                shutil.copy(config.custom_config, configFile)
+                print("Configuration backup restored!")
+            except:
+                print("Failed to apply custom backup!")
+        else:
+            backupFile = os.path.join(storageDir, "config_lite_backup.py" if config.isLite else "config_backup.py")
+            if os.path.isfile(backupFile):
+                restore_backup = yes_no_dialog(
+                    title="Configuration Backup Found",
+                    text=f"Do you want to use the following backup?\n{backupFile}"
+                ).run()
+                if restore_backup:
+                    try:
+                        loadConfig(backupFile)
+                        shutil.copy(backupFile, configFile)
+                        print("Configuration backup restored!")
+                        #config.restartApp()
+                    except:
+                        print("Failed to restore backup!")
 
 # load new / unsaved configs
 setConfig(defaultSettings)
